backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
python-decouple
psycopg2-binary
django-cors-headers
requests
gql
requests_toolbelt
channels
channels-redis
daphne
